+++ eslint.config.js
@@ -1,25 +1,20 @@
-import globals from 'globals'
-import js from '@eslint/js'
-
 export default [
-    js.configs.recommended,
     {
+        ignores: ['node_modules', 'public', 'dist', 'test/fixtures', 'test-temp'],
+    },
+    {
+        files: ['**/*.js'],
         languageOptions: {
-            globals: {
-                ...globals.node,
-                ...globals.es2021,
-            },
-            ecmaVersion: 2021,
+            ecmaVersion: 'latest',
             sourceType: 'module',
         },
         rules: {
-            'no-unused-vars': 'warn',
-            'no-console': 'off',
+            indent: ['error', 4],
+            semi: ['error', 'never'],
+            quotes: ['error', 'single'],
+            'no-unused-vars': ['warn'],
             'prefer-const': 'error',
             'no-var': 'error',
         },
-    },
-    {
-        ignores: ['node_modules/', 'dist/', 'public/', 'test/fixtures/'],
     },
 ]